Separate daemon/graphdriver/overlay/copy into its own package
